src/org/jessies/talc/Talc.java: add a couple of new debugging flags; one enables an incomplete bytecode compiler, the other dumps the generated bytecode before running it. For now the interpreter remains the default because it implements more of the language than the compiler. Also add a user-visible reference to the ASM library to the output of "talc --copyright".
src/org/jessies/talc/TalcClassLoader.java: a trivial class loader so we can load the classes we generate.

src/org/jessies/talc/SourceLocation.java: make the source filename available so we can embed it in the classes we generate (for the benefit of stack traces). We don't currently do anything about line numbers.

src/org/jessies/talc/Functions.java: the compiler won't use the same mechanism for built-in functions that the interpreter used, but I do like the name of this class, so for a while it will contain built-ins for the compiler and built-ins for the interpreter. At the moment, while the compiler's so incomplete, just having "puts" is adequate.

src/org/jessies/talc/JvmCodeGenerator.java: a partial implementation of a JVM bytecode compiler for Talc. Major limitations: no classes, methods/functions, loops of any kind, fields/variables.

tests/compiler-trip.talc: a simple test of those parts of the compiler that are currently implemented. Eventually we'll want some kind of automated test suite, but this will do for now, especially since Talc doesn't even have an "assert" statement yet.