src/org/jessies/talc/AstEvaluator.java: src/org/jessies/talc/Functions.java: fix == and != in the interpreter by farming out the logic to Functions.eq and Functions.ne (and using the correct logic there!).
src/org/jessies/talc/JvmCodeGenerator.java: fix the compiler in the same way. Also stop code in <init> overwriting our reference to "this" (which is in local slot 0), before that catches us out (and because it makes the disassembly harder to read). Finally, ensure that we can use variable definitions in code like "while ((s := gets()) != null) {".